Remove getAllRoles' given-condition workaround now that relocateConditions fixes it at the source
relocateConditions (added for #238) already keeps an inverted existential
condition off a given unless every label it references is the given itself,
so getAllRoles no longer needs its own label-seeding workaround for
existentials lifted onto a given that reference a sibling match label. That
shape can't come out of the inverter anymore.

Removed the hand-built regression test that pinned the workaround directly
(the shape it constructed is no longer producible) and instead assert
well-orderedness structurally on every inverse of the nested-existential
regression spec, sharing the check via a new specificationTestHelpers module
instead of duplicating it across two spec files.
