+++ src/specification/specification.ts
@@ -167,23 +167,15 @@ export function getAllRoles(specification: Specification): RoleDescription[] {
         }),
         {} as TypeByLabel);
 
-    // Walk the matches first so that the label map includes every unknown in
-    // the specification, not just the givens. An existential condition lifted
-    // onto a given (as inversion does) can path-reference a sibling match
-    // label; without those labels seeded, resolving its roles would fail with
-    // "Label <name> not found".
-    const { roles: rolesFromMatches, labels: labelsFromMatches } = getAllRolesFromMatches(labels, specification.matches);
-
-    // Collect roles from existential conditions on givens, using the full
-    // label map so sibling references resolve.
     let rolesFromGivenConditions: RoleDescription[] = [];
     for (const given of specification.given) {
         for (const condition of given.conditions) {
-            const { roles } = getAllRolesFromMatches(labelsFromMatches, condition.matches);
+            const { roles } = getAllRolesFromMatches(labels, condition.matches);
             rolesFromGivenConditions.push(...roles);
         }
     }
 
+    const { roles: rolesFromMatches, labels: labelsFromMatches } = getAllRolesFromMatches(labels, specification.matches);
     const components = specification.projection.type === "composite" ? specification.projection.components : [];
     const rolesFromComponents = getAllRolesFromComponents(labelsFromMatches, components);
     const roles: RoleDescription[] = [ ...rolesFromGivenConditions, ...rolesFromMatches, ...rolesFromComponents ];
