+++ test/specification/getAllRolesSpec.ts
@@ -1,55 +1,7 @@
-import { buildModel, getAllRoles, invertSpecification, Specification, User } from "@src";
+import { buildModel, getAllRoles, invertSpecification, User } from "@src";
+import { expectWellOrdered } from "./specificationTestHelpers";
 
 describe("getAllRoles", () => {
-    // Regression for jinaga/jinaga-replicator#52. When a positive existential
-    // (.exists) is inverted, the resulting inverse can carry that existential on
-    // its given, with an inner match that path-references a SIBLING match label
-    // rather than the given itself. The Postgres store resolves role IDs through
-    // getAllRoles, which previously seeded its label map with only the given
-    // labels and threw "Label <name> not found" on such inverses — surfacing as
-    // an HTTP 500 on /save. (The in-memory store never calls getAllRoles, which
-    // is why JinagaTest did not reproduce it.)
-    it("resolves an existential on the given that references a sibling match label", () => {
-        const specification: Specification = {
-            given: [{
-                label: { name: "u3", type: "Redemption" },
-                conditions: [{
-                    type: "existential",
-                    exists: true,
-                    matches: [{
-                        unknown: { name: "u4", type: "Receipt" },
-                        conditions: [{
-                            type: "path",
-                            rolesLeft: [{ name: "request", predecessorType: "Request" }],
-                            labelRight: "u2",
-                            rolesRight: []
-                        }]
-                    }]
-                }]
-            }],
-            matches: [{
-                unknown: { name: "u2", type: "Request" },
-                conditions: [{
-                    type: "path",
-                    rolesLeft: [],
-                    labelRight: "u3",
-                    rolesRight: [{ name: "request", predecessorType: "Request" }]
-                }]
-            }],
-            projection: { type: "fact", label: "u2" }
-        };
-
-        // Must not throw "Label u2 not found" — the sibling reference resolves.
-        const roles = getAllRoles(specification);
-        // The role referenced through the sibling label must be resolved with
-        // the correct successor type (Receipt -> request -> Request).
-        expect(roles).toContainEqual({
-            successorType: "Receipt",
-            name: "request",
-            predecessorType: "Request"
-        });
-    });
-
     it("resolves roles for every inverse of a spec with a nested positive existential", () => {
         class Domain { static Type = "Domain" as const; type = Domain.Type; constructor(public identifier: string) { } }
         class Provisioner { static Type = "Provisioner" as const; type = Provisioner.Type; constructor(public user: User, public domain: Domain) { } }
@@ -72,6 +24,7 @@ describe("getAllRoles", () => {
 
         const inverses = invertSpecification(specification.specification);
         for (const inverse of inverses) {
+            expectWellOrdered(inverse.inverseSpecification);
             expect(() => getAllRoles(inverse.inverseSpecification)).not.toThrow();
         }
     });
